requirements.txt
# Python 3.9.2
# Google Chrome 92.0.4515.159
# ChromeDriver 92.0.4515.107

et-xmlfile==1.1.0
openpyxl==3.0.7
Pillow==8.3.1
pytesseract==0.3.8
selenium==3.141.0
urllib3==1.26.6
